% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootnloot.R
\name{new_id}
\alias{new_id}
\title{Define new id}
\usage{
new_id(
  role = c("reader", "commenter", "writer", "fileOrganizer", "owner", "organizer"),
  type = c("user", "group", "domain", "anyone"),
  ...,
  name = "shoot.loot"
)
}
\arguments{
\item{role}{Character. The role to grant. Must be one of:
\itemize{
\item owner (not allowed in shared drives)
\item organizer (applies to shared drives)
\item fileOrganizer (applies to shared drives)
\item writer
\item commenter
\item reader
}}

\item{type}{Character. Describes the grantee. Must be one of:
\itemize{
\item user
\item group
\item domain
\item anyone
}}

\item{...}{Name-value pairs to add to the API request. This is where you
provide additional information, such as the \code{emailAddress} (when grantee
\code{type} is \code{"group"} or \code{"user"}) or the \code{domain} (when grantee type is
\code{"domain"}). Read the API docs linked below for more details.}

\item{name}{The name of the file on google drive, does not not impact
the use of the package}
}
\value{
An id
}
\description{
This uploads an empty file to Google drive and sets permissions so the file
can then be updated by \code{shoot()} and downloaded with \code{loot()}.
The id should then be set into \code{options(shootnloot.ids = c(your_label = "your_new_id"))}.
}
\details{
See examples in \code{?shoot}
}
