% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootnloot.R
\name{shootnloot}
\alias{shootnloot}
\alias{shoot_object}
\alias{loot_object}
\alias{shoot_file}
\alias{loot_file}
\alias{shoot_workspace}
\alias{loot_workspace}
\title{Shoot and loot}
\usage{
shoot_object(obj, where = "default")

loot_object(where = "default")

shoot_file(path = NULL, where = "default")

loot_file(where = "default")

shoot_workspace(where = "default")

loot_workspace(where = "default")
}
\arguments{
\item{obj}{An object to upload}

\item{where}{The label of the id to upload (shoot) to or download (loot) from. You'll need the appropriate rights.}

\item{path}{Path to a file to upload, if left \code{NULL}, the active file
(including unsaved changes) is uploaded}
}
\value{
\code{shoot_object()} and \code{loot_object()} return the object, \code{shoot_object()} returns it invisibly since it's
called for side effects. \code{shoot_file()}, \code{loot_file()}, \code{shoot_worspace()} and \code{loot_workspace()} return \code{NULL} invisibly
since they are called for side effects.
}
\description{
Upload or download an object, a script, or the entire global environment
}
